feat: 회원가입 폼 관련 타입 및 초기값을 알맞은 폴더로 나눠서 저장
