+++ src/constants/signupForm.constants.ts
@@ -1,0 +1,12 @@
+// 회원가입 폼 에러 메시지 상수
+export const SIGNUP_ERROR_MSG = {
+  name: '이름을 입력하세요.',
+  email: '이메일을 입력하세요.',
+  emailFormat: '올바른 이메일 형식을 입력하세요.',
+  phone: '전화번호는 000-0000-0000 형식이어야 합니다.',
+  password: '비밀번호를 입력하세요.',
+  passwordCheck: '비밀번호가 일치하지 않습니다.',
+  emailSend: '이메일 인증코드 전송 실패',
+  emailNetwork: '네트워크 오류',
+  emailCode: '인증코드가 올바르지 않습니다'
+};
