+++ src/hooks/useSignupForm.ts
@@ -1,46 +1,15 @@
 import { useState } from 'react';
 
-// 회원가입 폼 값 타입
-export interface SignupValues {
-  name: string;
-  phone: string;
-  email: string;
-  emailCode: string;
-  password: string;
-  passwordCheck: string;
+import { SIGNUP_ERROR_MSG } from '../constants/signupForm.constants';
+import type { SignupFormError, SignupValues } from '../types/signupForm.types';
+
+function isValidPhone(phone: string) {
+  return /^\d{3}-\d{3,4}-\d{4}$/.test(phone);
+}
+function isValidEmail(email: string) {
+  return /^[\w-.]+@[\w-]+\.[a-zA-Z]{2,}$/.test(email);
 }
 
-// 회원가입 폼 초기값
-export const signupInitialForm: SignupValues = {
-  name: '',
-  phone: '',
-  email: '',
-  emailCode: '',
-  password: '',
-  passwordCheck: ''
-};
-
-// 에러 타입
-export type SignupFormError = Partial<
-  Record<'name' | 'phone' | 'email' | 'emailCode' | 'password' | 'passwordCheck', string>
-> & {
-  emailAuth?: string;
-};
-
-// 에러 메시지 상수
-export const SIGNUP_ERROR_MSG = {
-  name: '이름을 입력하세요.',
-  email: '이메일을 입력하세요.',
-  emailFormat: '올바른 이메일 형식을 입력하세요.',
-  phone: '전화번호는 000-0000-0000 형식이어야 합니다.',
-  password: '비밀번호를 입력하세요.',
-  passwordCheck: '비밀번호가 일치하지 않습니다.',
-  emailSend: '이메일 인증코드 전송 실패',
-  emailNetwork: '네트워크 오류',
-  emailCode: '인증코드가 올바르지 않습니다'
-};
-
-// 커스텀 훅
 export function useSignupForm(initialForm: SignupValues, onSubmit: (values: SignupValues) => void) {
   const [form, setForm] = useState<SignupValues>(initialForm);
   const [error, setError] = useState<SignupFormError>({});
@@ -48,12 +17,6 @@ export function useSignupForm(initialForm: SignupValues, onSubmit: (values: Sign
   const [emailSent, setEmailSent] = useState(false);
   const [emailVerified, setEmailVerified] = useState(false);
 
-  function isValidPhone(phone: string) {
-    return /^\d{3}-\d{3,4}-\d{4}$/.test(phone);
-  }
-  function isValidEmail(email: string) {
-    return /^[\w-.]+@[\w-]+\.[a-zA-Z]{2,}$/.test(email);
-  }
   function clearFieldError(name: string) {
     setError((prev) => ({ ...prev, [name]: undefined }));
   }
